osd: optimize handling of RGW's start_after during OMAP iteration
Currently an OSD does up to 3 key seeks while iterating over OMAP:
1. to rewind to the first OMAP entry of particular RADOS object
   in the whole, global key namespace;
2. then to find the pagination-related `start_after` key;
3. then to respect the `filter` parameter.

This is pretty inefficient as `Seek()` in RocksDB is logarithmic,
so `Seek(n)` should be cheaper than `Seek(n/2)` + `Seek(n/2)`.

Please note there are other inefficiencies. The underlying `Seeks()`
are performed under the collection lock and the iterator is one-time
only. Ultimately we could try to cache it and reuse across multiple
rounds of the same listing of RGW.

Fixes: https://tracker.ceph.com/issues/68457

Signed-off-by: Radoslaw Zarzynski <rzarzyns@redhat.com>