test/osd: Add interactive mode to ceph_test_rados_io_sequence
Reorganisers the main function to be more modular, moving functionality into a TestRunner object and anonymous namespaces.
Adds a new "interactive" test mode that allows you to direct the order of operations rather than using the preset sequences. This allows complete precise control of the tool when you want to test specific IOs.

Signed-off-by: Jon Bailey <jonathan.bailey1@ibm.com>