rbd-mirror: mirror clone after parent snap data copy
The clone image is mirrored to the non-primary before the parent snap
data has been copied making it possible to get incorrect data on
accessing the non-primary clone image.
The fix checks that the parent image snap from which the clone was created
has been synced before creating the non-primary clone image.

Fixes: https://tracker.ceph.com/issues/68205

Signed-off-by: N Balachandran <nibalach@redhat.com>