rgw: cleanup special handling of errors
This changes the rgw_http_errors from being a pair
to a tuple that now also includes a potential default
error message.

This removes the special cases added in the code just
to be able to set an error message in a good way.

This changes the error message for an expired presigned
URL to "Request has expired" as per the AWS S3 [1]
standard message.

A future improvement here is addressing the comment about
cleaning up the error handling in Swift, another improvement
is getting rid of the set_req_state_err() signature that
takes in a error message and rework that to something like
set_req_state_err_msg and rely on everything else by default
from rgw_http_errors.

[1] https://repost.aws/knowledge-center/request-has-expired-s3-object

Signed-off-by: Tobias Urdin <tobias.urdin@binero.com>