mgr/dashboard: improve the kcli bootstrap process
I have a new script added for starting the kcli cluster called
quick-bootstrap.sh

The goal is to use that script to download the ceph image on local
(rather than inside vm) and then copy them over to all the vms that is
being spawned by the kcli. This way all the hosts will get the ceph
image which will make the deployment loads faster.

Another thing I added is to add some dnf.conf to improve
parallel_downlaods and get the fastest server to install deps

Signed-off-by: Nizamudeen A <nia@redhat.com>