qa: Run test_admin with the squid client
This is required to test the features involving
fixes both in client and mds. This is to make
sure the older clients are not broken with the
fix. The version 19.2.2 is used for client.

The test suite sets up the cluster with squid
19.2.2 and upgrades only the ceph cluster node
leaving the client node.

Signed-off-by: Kotresh HR <khiremat@redhat.com>
(cherry picked from commit a4f97c0aa92c37113b33d63b57f2fae870f403a1)