qa: Run test_admin with the reef client
This is required to test the features involving
fixes both in client and mds. This is to make
sure the older clients are not broken with the
fix. The version 18.2.6 is reef with out the
client fix.

The test suite sets up the cluster with squid
18.2.6 and upgrades only the ceph cluster node
leaving the client node.

NOTE: The version is changed to 18.2.6 because
this is reef backport where as it's 19.2.2 in
higher releases. Please check commit a4f97c0aa92

Signed-off-by: Kotresh HR <khiremat@redhat.com>
(cherry picked from commit a4f97c0aa92c37113b33d63b57f2fae870f403a1)