mgr/cephadm: add offline host handling to maintenance exit
The maintenance exit command now will check if the host
is offline before attempting to take it out of maintenance
mode (previously this case would cause the command to fail
and print a not very useful traceback). Additionally, if users
would like to have cephadm officially consider an offline host to
be no longer in maintenance mode this patch allows users to
pass a --offline and --force flag to do so. The expected use
case for those flags being a host that is considered to be
permanently online that users want to mark as no longer being
in maintenance mode before the host is removed from the cluster.

Signed-off-by: Adam King <adking@redhat.com>
(cherry picked from commit ee9b7c0617c972dcdba15902c280baebe001504c)