cls/timeindex: make ext key unique
I.e. when adding an entry with key_ext that already exists the old
entry will be removed.

Although it changes expectations about how timeindex class works,
the only user of the class is rgw and it wants this behavior:
when adding an index (an object expiration time) we want the old index with
the same ext key (object name) to be removed as it is not valid anymore.

Fixes: https://tracker.ceph.com/issues/68380
Signed-off-by: Mykola Golub <mykola.golub@clyso.com>