dencoding: check struct_v against DECODE_START(v, ...) at compile-time
Typical problem found in recent reviews of dencoding changes is missed
increment of the version number passed to `DECODE_START()` macro in
a decoder.
It's easy to overlook and hard to find as the number is used for rarely
happening explicit breaks of schema compatibility.

This commit brings compile-time checks for `struct_v` in decoders
(and for `denc` also in encoders) to find a situation when it's
compared against version higher than declared using the `DECODE_START(v, ...)`
macro.

Additionally, the patch fixes the currently existing issues of this
genre buried across the enitre code base.

Signed-off-by: Radoslaw Zarzynski <rzarzyns@redhat.com>