src/mon/PGMap.cc: check `degraded` state in `get_unavailable_pg_in_pool_map`
If a pool has any PG with unfound objects, we should consider
it unavailable for the availability score. If a PG has unfound
objects, it will have a `degraded` state.

In `get_unavailable_pg_in_map`, if a PG's degraded state bit is
set, we add it to `pool_pg_unavailable_map`.

Signed-off-by: Shraddha Agrawal <shraddha.agrawal000@gmail.com>