doc/governance: revert d7c144c0d7e490bfda9715a0c8462bb67cffc764
Revert the commit (manually, by restoring the file by hand to the state
it was in prior to d7c144c) to the state that it was in before I added
the Executive Council Responsibilities document to governance.rst. This
document cannot be edited at will, but must be voted on by the
Leadership Team.

Signed-off-by: Zac Dover <zac.dover@proton.me>