cephamd: add status details for init and sidecar services
Add status information to the previous set of init container service and
sidecar container services similar to the basic unit status provided for
the primary (default) services. When a daemon includes either an init
or sidecar service the 'status' key will become an aggregate status -
summarizing the status of all container services making up the daemon.
The non-aggregate state of the individual (primary) unit is then included
in the dictionary at the 'unit_state' key.

Signed-off-by: John Mulligan <jmulligan@redhat.com>