common/fs_types: combine several fixed-size `encode()` calls
The `denc` library allows appending multiple values in a single
`buffer::list` call by wrapping them in a `std::tuple`.  This reduces
overhead because the buffer bounds checks have to be performed only
once.

The `file_layout_t` type turned up in the profiler, so it was an
obvious choice to optimize now.

Signed-off-by: Max Kellermann <max.kellermann@ionos.com>