rgw/d4n: squashing following commits related to lazy deletion - delete dirty objects data blocks in cleaning method and non-dirty objects data blocks in eviction method
1. rgw/d4n: Implement lazy deletion
2. rgw/d4n: cleaning method now supports delete markers
also (both versioned and null). The delete markers
are written correctly to the backend store.
Also modifying the description of rgw_d4n_cache_cleaning_interval
to explicitly state that the duration is in seconds.
3. rgw/d4n: do not call invalidate_dirty_object in case
of a simple delete request for dirty objects belonging
to a versioned bucket.
In this case, a delete marker needs to be created instead of
invalidating/deleting an object.
4. rgw/d4n: Update lazy deletion in policy

Co-authored-by: Pritha Srivastava <prsrivas@redhat.com>
Added code for creation of delete marker for a simple
delete request.

Signed-off-by: Samarah <samarah.uriarte@ibm.com>
Signed-off-by: Pritha Srivastava <prsrivas@redhat.com>

Commit 2c1adbbd2b363cd5e02fd7da0cb496ca6a93aa77

d4n/policy.cc: Use ceph::split

Signed-off-by: Samarah <samarah.uriarte@ibm.com>