osd/recovery_types: Introduce BackfillInterval::update
Updating the interval should be BackfillInterval
responsibilty only.

*** This section needs verification ***

Beside moving to logic over to BackfillInterval::update,
we now also update the version on each update() call.

When updating the ragne of the interval, we first scan the pg log
up to it's latest entry and later on scan the projected log as well.

With this chnage, we will now scan the projected log *only* from
pg_log's latest entry up to the projected_log latest entry.
Since after finishing to scan the pg log, the version will
be updated correspondly.

This will result in BackfillIntervall::vesion to be the
projeceted_log latest entry. Instead of previously only setting
the version to projected_last_update (after scanning both logs).

Note: Is it safe to assume that:
get_projected_last_update == projeceted_log last entry version

***************************************

Signed-off-by: Matan Breizman <mbreizma@redhat.com>