doc/cephfs: edit "Disabling Volumes Plugin"
Improve the English language in the section 'Disabling Volumes Plugin"
(which here becomes "Disabling the Volumes Plugin") in the
doc/cephfs/troubleshooting.rst file.

The change here is similar to a change introduced in
https://github.com/ceph/ceph/pull/60448/. That PR seems to be unable to
pass the "make check", and this PR is part of a test to see if breaking
the changes in that PR into several smaller changes results in mergeable
PRs.

Signed-off-by: Zac Dover <zac.dover@proton.me>
(cherry picked from commit 3e043a903ca26ce153093cbd7c498e2577c41d8a)