cephadm: use structured class to build systemd drop-in files
This is a precursor to both adding a special case drop-in for
node-exporter and for having custom systemd drop-ins specified
in a service spec for daemons of that service. I felt previously
the drop-in related code we had was very specific to init and
sidecar containers, whereas this makes it more generic.

Signed-off-by: Adam King <adking@redhat.com>