doc/cephfs: backport "Disabling Volumes Plugin"
Backport the "Disabling Volumes Plugin" section of
doc/cephfs/fs-volumes.rst, which was folded into a packet of code
changes in https://github.com/ceph/ceph/pull/58647 and therefore not
possible to backport in the normal way. This is one of a series of
similar pull requests that will result in this file being congruent
across release branches.

Signed-off-by: Zac Dover <zac.dover@proton.me>