mon/OSDMonitor: set default osd_beacon_report_interval after election
The last_osd_report vector is empty right after a new election of the monitor.
This causes the monitor to check for OSD beacons with
osd_beacon_report_interval = 0.
If mon_osd_report_timeout is smaller than osd_beacon_report_interval,
we will not wait long enough for some of the OSDs' beacons,
which can cause them to be marked as down.

The fix is to use the default osd_beacon_report_interval, which will make
the beacon check more reasonable and prevent OSDs from being marked
as down without any reason.

Fixes: https://tracker.ceph.com/issues/68363
Signed-off-by: Nitzan Mordechai <nmordech@redhat.com>