qa/cephadm: add timeouts to checks that daemons have been upgraded
There was another patch earlier that did this for a specific check
because we saw `ceph orch ps` reporting an RGW daemon as upgraded
faster than `ceph versions` did. This patch extends this to all
our current upgrade checks. The hope is it makes the tests more
consistent and we no longer have to worry about small timing
differences between what cephadm knows and the ceph cluster knows.

This should probably also be done in a more controlled way than
just having to run "time timeout 60..." in a whole bunch of differnet
spots but this is at least a start and allows us to see if it
makes any notable differece

Signed-off-by: Adam King <adking@redhat.com>