dbstore: add StripeNum to objdata_table and populate
This is a flag day for objdata_table* schema, and the meaning of
partnum is now always the upload part ordinal for multipart upload
objects, or 0 if not mpu.

Additional tail chunks are now indexed from 0 as StripeNum in the same
table (formerly this value was stored as PartNum, and upload part
number could be recovered only by parsing multipart_part_str).

Signed-off-by: Matt Benjamin <mbenjamin@redhat.com>