rgw: hardcode zones in sync pipes for specified buckets
With the introduction of replication between zonegroups, the replication
flow typically includes all zones across all zonegroups. This results in
inheriting the flow for individual buckets, which poses a problem: a
bucket can not be replicable to all zones in every zonegroup.

To address this, we now hardcode the zones in sync pipes when '*' is
passed as the dest_zones. However, we retain a hint that indicates
the user intended the bucket to be replicated wherever it resides.

This change repurposes the all_zones field in rgw_sync_bucket_entities to
act as a hint for this intent, instead of signaling a literal '*'. The
actual '*' (replicating to all zones) is now represented by nullopt for
zones in cases where buckets are not involved, such as base zonegroup
policies.

Additionally, to ensure the sync pipes remain valid when a bucket is
recreated, RGWCreateBucket now checks for any existing hints tied to
the bucket and updates the pipes accordingly. This guarantees that
pipes continue to function even if the bucket is recreated in a
different zonegroup.

Signed-off-by: Seena Fallah <seenafallah@gmail.com>