rgw:dbstore: use UpdateObjectData to set ObjInstance and Mtime
Fix assignment of the ObjInstance column in objectdata_table during
complete-multipart.

The effective change is to adapt an existing update of Mtime for these
rows to also update ObjInstance with the instance id (if any), which for
multipart uploads is only generated when an upload completes.

Minor changes, which may require cleanup, included adding an UploadID
column tracking the multipart upload id.  It now looks like this was
already being tracked in "ObjID."  We probably only need one of these.
It's not clear why the original name was chosen, as it is confusing
to the reader.

Signed-off-by: Matt Benjamin <mbenjamin@redhat.com>