rgw: introduce NotApplicable bucket sync state
This bucket sync state used for buckets that cannot be replicated
to the zonegroup or they have been deleted in the middle of the
sync and so we don't want to keep the bilogs so the zonegroup needs
to report a bilog status or them as the pipe for the source bucket
still points to that zone.
On bilog trimming, the peers that got this status will be excluded
from min generation and min position calculation as they are not
intrested in replicating that bucket.

Signed-off-by: Seena Fallah <seenafallah@gmail.com>