rgw: report last processed marker in datalog list
With the introduction of log_zones, a zone might not receive certain
datalog entries if they are irrelevant to that zone. To support
proper trimming, we now return the last processed marker, allowing
zones to report this marker even when they haven't processed those
irrelevant entries. This ensures that the source zone can proceed
with trimming.

Signed-off-by: Seena Fallah <seenafallah@gmail.com>