rgw: handle EINVAL translation in forward_request
RGWRESTSimpleRequest::forward_request() translates EINVAL to a 503
status code to reflect connection errors as internal errors rather
than client errors. However, if the master is explicitly returning
EINVAL (which corresponds to an HTTP 400 status code), this
translation incorrectly converts it to 503, hiding the true client
error.

To address this, we now check the HTTP status code to ensure the
error originates from easy_curl and not from a valid HTTP response,
preserving the intended 400 status when appropriate.

Fixes: https://tracker.ceph.com/issues/69084
Signed-off-by: Seena Fallah <seenafallah@gmail.com>