mds: Fix remote/referent journalling and replay
In multi-version inode, i.e., if a file has hardlinks
and the primary link is being deleted, the primary
inode is added as remote in the journal while journaling
cow. This has to be remote since it doesn't have a
corresponding referent inode.

Fix the journal to handle the same.

Signed-off-by: Kotresh HR <khiremat@redhat.com>