qa/tasks/nvme_loop: update task to work with new nvme list format
Specifically on some centos 9 tests, we've seen that a newer
version of some nvme related package is causing this task to fail
with "KeyError: 'DevicePath'" due to the format of the output
of the nvme list command changing. This patch adds handling for
the new format we've seen while also still supporting the old
format (necessary for the tests running on ubuntu).

Fixes: https://tracker.ceph.com/issues/69067

Signed-off-by: Adam King <adking@redhat.com>
(cherry picked from commit ce487613d08b0c0d73172ce6ababcb75855f6ae6)