qa/workunits: add new collections to telemetry upgrade workunit (reef-->x)
This commit changes the workunit that tests the upgrade path from
reef to x.

The logic to verify collections is now much more robust. Previously,
the test relied on developers to make it aware of any new collections
it needed to verify. Now, I have added logic to check that the # of
known collections matches the # of reported collections. If it does not,
the test errors out, which will force developers to add any new
collections to the upgrade test along with adding it to the telemetry
code.

The test also verifies all collections instead of focusing only on new
collections that were added after the base collections.

Signed-off-by: Laura Flores <lflores@ibm.com>