crimson/osd/recovery_backend: scan_for_backill tries to load the metadata even if the obc is already in the cache while its existed is false
Currently there are 3 cases (before the patch):

obc found in cache && exists is true -> emplace it.
obc found in cache && exists is false -> return
obc not found -> load_metadata

We should reload the metadata **also** when exists is false in order to
exclude the possibility that there's an ongoing obc load.

Fixes: https://tracker.ceph.com/issues/69154
Signed-off-by: Xuehan Xu <xuxuehan@qianxin.com>