rgw-admin: allow switching between sync states for bucket
As part of the recent update in https://github.com/ceph/ceph/pull/61003,
a new command `bucket sync batch-replicate` has been introduced to allow
a bucket to perform batch replication when it is initially set to start
with incremental replication.

Additionally, a `bucket sync skip-full-sync` command has been introduced,
enabling the switch from full sync to incremental sync. However, this
command applies only to buckets where the source and destination names
differ, as symmetry must be maintained when the names are the same.

Fixes: https://tracker.ceph.com/issues/69309
Signed-off-by: Seena Fallah <seenafallah@gmail.com>