mgr/vol: make clone progress reporter wait before termination
Class CloneProgressReporter spawns a thread that does the work of
tracking ongoing and pending clone jobs and producing statistics for
them. This thread is terminated as soon as all clone jobs finish.

Instead of terminating the thread and releasing the related resources
right away, wait for new clones to arrive for a minute or so. Thus if a
new clones arrives during the waiting period, the thread and related
resources are reused instead of throwing away and reallocating again.

Fixes: https://tracker.ceph.com/issues/69206
Signed-off-by: Rishabh Dave <ridave@redhat.com>