mgr/cephadm: change exceptions received from purging services into OrchestratorError
The cephadm serve loop has handling that can take an OrchestratorError, log
an event that users can view, and then continue on with another iteration of
the serve loop. Other exception types are not caught. This commit makes sure of
two things. Firstly, it makes sure exceptions from purging services are always
caught by the serve loop to avoid unhandled exception types causing the cephadm
module to crash. Second, it raises an OrchestratorError instead of just logging
the issue so that we do not go on to fully remove the spec from the spec store.
This makes it so that the purge action will be attempted again later, rather
than continuing on and hoping the purge action wasn't necessary.

Signed-off-by: Adam King <adking@redhat.com>